+++ data-plane/receiver/src/main/java/dev/knative/eventing/kafka/broker/receiver/impl/handler/MethodNotAllowedHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.receiver.handler;
+package dev.knative.eventing.kafka.broker.receiver.impl.handler;
 
 import io.vertx.core.Handler;
 import io.vertx.core.http.HttpMethod;
